Make Geocoding testable from desktop unit test projects
This showcases how a minimal change in the netstandard version of the
API can trivially allow for fully testable API invocations without
incurring in *any* of the run-time performance compromises of switching
the platform implementations to use either unnecessary interface/virtual
calls or even instance method call overhead ('cause of the null reference
check that's otherwise not necessary).

This is an effective compromise that covers the majority of the testing
scenarios (unit testing from view models down) while still preserving
the desired run-time characteristics of the project.