Allow parallel tests to set their own mocks
By storing the Current instance in AsyncLocal, it can be
properly set per-thread and propagated automatically across
async calls.

NOTE: since AsyncLocal<T> only exists in netstandard2.0, I
just removed netstandard1.0 ;). Of course, this is just to
show how we could do this, but if we support this for real,
we'd need to have different implementations for ns1.0 vs
ns2.0.